+++ src/pages/ShoppingListAnimeGangRedux/index.tsx
@@ -10,13 +10,13 @@ import {
   TopRow,
 } from './styles/ShoppingListAnimeGangRedux.styled'
 import { RootState } from '../../store/types'
-import { ShoppingListItem } from '../../store/AnimeGangRedux/types'
-import { AddQuantityToItem, DecreaseQuantityToItem, RemoveItemFromList } from '../../store/AnimeGangRedux/actions'
+import { ShoppingListItem } from '../../store/animeGangRedux/types'
+import { AddQuantityToItem, DecreaseQuantityToItem, RemoveItemFromList } from '../../store/animeGangRedux/actions'
 import { useDispatch, useSelector } from 'react-redux'
 
 const ShoppingAnimeGangRedux = () => {
   const dispatch = useDispatch()
-  const { shoppingList } = useSelector((state: RootState) => state.AnimeGangRedux)
+  const { shoppingList } = useSelector((state: RootState) => state.animeGangRedux)
   const columns = [
     {
       title: 'Quantity',
@@ -35,7 +35,11 @@ const ShoppingAnimeGangRedux = () => {
       render: (text: ShoppingListItem, record: ShoppingListItem, index: number) => (
         <Space size="middle">
           <Button onClick={() => dispatch(AddQuantityToItem(index))} shape="circle" icon={<PlusOutlined />}></Button>
-          <Button onClick={() => dispatch(DecreaseQuantityToItem(index))} shape="circle" icon={<MinusOutlined />}></Button>
+          <Button
+            onClick={() => dispatch(DecreaseQuantityToItem(index))}
+            shape="circle"
+            icon={<MinusOutlined />}
+          ></Button>
           <a onClick={() => dispatch(RemoveItemFromList(index))}>Delete</a>
         </Space>
       ),
@@ -46,9 +50,9 @@ const ShoppingAnimeGangRedux = () => {
     <FullScreenContainer>
       <TopRow>
         <ShoppingListHeader> Shopping list</ShoppingListHeader>
-        <AddItemsContainer>
+        {/* <AddItemsContainer>
           {showAddItem && (
-            <Button onClick={() => handleOnAddItemClick()} type="primary">
+            <Button onClick={() => ShowAddItemInput()} type="primary">
               + Add item
             </Button>
           )}
@@ -59,10 +63,10 @@ const ShoppingAnimeGangRedux = () => {
               size="small"
               placeholder={addItemName}
               onChange={(e) => handleOnType(e.target.value)}
-              onPressEnter={() => handleOnEnter()}
+              onPressEnter={() => AddItemToList()}
             />
           )}
-        </AddItemsContainer>
+        </AddItemsContainer> */}
       </TopRow>
       <TablePart>
         <Table columns={columns} dataSource={shoppingList}></Table>
@@ -72,10 +76,3 @@ const ShoppingAnimeGangRedux = () => {
 }
 
 export default ShoppingAnimeGangRedux
-function reactRedux.useDispatch() {
-  throw new Error('Function not implemented.')
-}
-
-function reactRedux.useSelector(arg0: (state: RootState) => any): { shoppingList: any } {
-  throw new Error('Function not implemented.')
-}
