+++ src/store/store.ts
@@ -2,11 +2,12 @@ import { createStore, applyMiddleware, combineReducers } from 'redux'
 import thunk from 'redux-thunk'
 import { composeWithDevTools } from 'redux-devtools-extension'
 
-//import { shoppingLists } from './shoppingLists/reducer'
-import { AnimeGangRedux } from './AnimeGangRedux/reducer'
+import { shoppingLists } from './shoppingLists/reducer'
+import { animeGangRedux } from './animeGangRedux/reducer'
 
 export const rootReducer = combineReducers({
-  AnimeGangRedux,
+  animeGangRedux,
+  shoppingLists,
 })
 const middlewares = [thunk]
 const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(...middlewares)))
